+++ aqueduct/lib/src/http/response.dart
@@ -42,7 +42,7 @@ class Response implements RequestOrResponse {
   ///
   /// Where [lastModified] is the last modified date of the resource
   /// and [cachePolicy] is the same policy as applied when this resource was first fetched.
-  Response.notModified(DateTime lastModified, HTTPCachePolicy cachePolicy) {
+  Response.notModified(DateTime lastModified, CachePolicy cachePolicy) {
     statusCode = HttpStatus.notModified;
     headers = {HttpHeaders.lastModifiedHeader: HttpDate.format(lastModified)};
     this.cachePolicy = cachePolicy;
@@ -84,19 +84,19 @@ class Response implements RequestOrResponse {
 
   /// An object representing the body of the [Response], which will be encoded when used to [Request.respond].
   ///
-  /// This is typically a map or list of maps that will be encoded to JSON. If the [body] was previously set with a [HTTPSerializable] object
-  /// or a list of [HTTPSerializable] objects, this property will be the already serialized (but not encoded) body.
+  /// This is typically a map or list of maps that will be encoded to JSON. If the [body] was previously set with a [Serializable] object
+  /// or a list of [Serializable] objects, this property will be the already serialized (but not encoded) body.
   dynamic get body => _body;
 
   /// Sets the unencoded response body.
   ///
-  /// This may be any value that can be encoded into an HTTP response body. If this value is a [HTTPSerializable] or a [List] of [HTTPSerializable],
-  /// each instance of [HTTPSerializable] will transformed via its [HTTPSerializable.asMap] method before being set.
+  /// This may be any value that can be encoded into an HTTP response body. If this value is a [Serializable] or a [List] of [Serializable],
+  /// each instance of [Serializable] will transformed via its [Serializable.asMap] method before being set.
   set body(dynamic initialResponseBody) {
     dynamic serializedBody;
-    if (initialResponseBody is HTTPSerializable) {
+    if (initialResponseBody is Serializable) {
       serializedBody = initialResponseBody.asMap();
-    } else if (initialResponseBody is List<HTTPSerializable>) {
+    } else if (initialResponseBody is List<Serializable>) {
       serializedBody =
           initialResponseBody.map((value) => value.asMap()).toList();
     }
@@ -139,9 +139,9 @@ class Response implements RequestOrResponse {
   /// Cache policy that sets 'Cache-Control' headers for this instance.
   ///
   /// If null (the default), no 'Cache-Control' headers are applied. Otherwise,
-  /// the value returned by [HTTPCachePolicy.headerValue] will be applied to this instance for the header name
+  /// the value returned by [CachePolicy.headerValue] will be applied to this instance for the header name
   /// 'Cache-Control'.
-  HTTPCachePolicy cachePolicy;
+  CachePolicy cachePolicy;
 
   /// The content type of the body of this response.
   ///