+++ aqueduct/test/http/cache_policy_test.dart
@@ -5,7 +5,7 @@ import 'package:http/http.dart' as http;
 import 'package:aqueduct/aqueduct.dart';
 import 'package:test/test.dart';
 
-// Some HTTPCachePolicy fields are tested by file_controller_test.dart, this
+// Some CachePolicy fields are tested by file_controller_test.dart, this
 // file tests the combinations not tested there.
 void main() {
   HttpServer server;
@@ -15,14 +15,14 @@ void main() {
   });
 
   test("Prevent intermediate caching", () async {
-    var policy = const HTTPCachePolicy(preventIntermediateProxyCaching: true);
+    var policy = const CachePolicy(preventIntermediateProxyCaching: true);
     server = await bindAndRespondWith(Response.ok("foo")..cachePolicy = policy);
     var result = await http.get("http://localhost:8888/");
     expect(result.headers["cache-control"], "private");
   });
 
   test("Prevent caching altogether", () async {
-    var policy = const HTTPCachePolicy(preventCaching: true);
+    var policy = const CachePolicy(preventCaching: true);
     server = await bindAndRespondWith(Response.ok("foo")..cachePolicy = policy);
     var result = await http.get("http://localhost:8888/");
     expect(result.headers["cache-control"], "no-cache, no-store");